Add files via upload
INTER-VLAN is configured using a (3560) layer 3 switch which works as a switch, 
router and also a DHCP server which means it is capable of multitasking and such 
can be perfect for small or a medium size business who wants to save cost and still 
keep an active and modern device for their network environment. It has a 3 layer 2 
switches which can also convert into a layer 3 interface to serve purpose of routing. 
The 3 layer switch will route through 3 (2960) switch, two will be connected with 
2 department/VLAN each and one will be connected with 2 department/VLAN and 
the server on a separate VLAN. 7 VLANs will be created in this Lab which will consist 
of a server VLAN and six departments with two computers each. The configuration 
with consist of VLANs, assigning access ports, configuring Trunking, Assigning IP/IP 
routing, configuring DHCP. The project will be done using Cisco Packet Tracer.